pyproject.toml
[tool.poetry]
name = "type-yaml"
version = "0.1.0"
description = ""
authors = ["nahco314 <nahco3_ta@yahoo.co.jp>"]
readme = "README.md"
packages = [{include = "type_yaml"}]

[tool.poetry.dependencies]
python = "^3.8.1"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
types-pyyaml = "^6.0.12.2"
black = "^22.12.0"
isort = "^5.11.4"
mypy = "^0.991"
pyproject-flake8 = "^6.0.0.post1"
tox = "^4.0.19"
coverage = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
line_length = 88
force_single_line = true

[tool.flake8]
max-line-length = 88
ignore = ["E203", "W503"]

[tool.mypy]
python_version = 3.8
disallow_untyped_defs = true

[tool.coverage.run]
relative_files = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\bProtocol\\):",
    '@(abc\.)?abstractmethod',
]

[tool.tox]
legacy_tox_ini = """
[tox]
env_list =
    py311
    py310
    py39
    py38

[testenv]
allowlist_externals = poetry
skip_install = true
commands =
    poetry install -v
    poetry run pytest
"""
